Move preparePathForShell into terminalEnvironment
