+++ src/vs/workbench/contrib/terminal/browser/terminalActions.ts
@@ -49,6 +49,7 @@ import { isAbsolute } from 'vs/base/common/path';
 import { ITerminalQuickPickItem } from 'vs/workbench/contrib/terminal/browser/terminalProfileQuickpick';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
 import { getIconId, getColorClass, getUriClasses } from 'vs/workbench/contrib/terminal/browser/terminalIcon';
+import { preparePathForShell } from 'vs/platform/terminal/common/terminalEnvironment';
 
 // allow-any-unicode-next-line
 export const switchTerminalActionViewItemSeparator = '─────────';
@@ -534,7 +535,7 @@ export function registerTerminalActions() {
 			}
 
 			// TODO: Convert this to ctrl+c, ctrl+v for pwsh?
-			const path = await terminalInstanceService.preparePathForTerminalAsync(uri.fsPath, instance.shellLaunchConfig.executable, instance.title, instance.shellType, instance.remoteAuthority);
+			const path = await preparePathForShell(uri.fsPath, instance.shellLaunchConfig.executable, instance.title, instance.shellType, (e) => terminalInstanceService.getBackend(instance!.remoteAuthority)?.getWslPath(e));
 			instance.sendText(path, true);
 			return terminalGroupService.showPanel();
 		}
